+++ lib/authrocket/jwt_key.rb
@@ -1,0 +1,11 @@
+module AuthRocket
+  class JwtKey < Resource
+    crud :all, :find, :create, :delete
+
+    belongs_to :realm
+
+    attr :algo, :key, :use
+    attr :expired # readonly
+
+  end
+end